requirements.txt
streamlit>=1.45.0
agno>=1.5.4
pandas>=2.2.0
pydantic>=2.11.0
python-dotenv>=1.0.0
praw>=7.8.0
yfinance>=0.2.0
duckduckgo-search>=8.0.0
openai>=1.82.0
google-generativeai>=1.16.0 